Homogenize a map from the Skip module.
I'd assumed I could use Homogenize to merge a mapping result from the
Skip module, but Homogenize expects the results to be sorted.  With
partial a set ordered `[ 'b', 'a' ]` the key `[ 'b' ]` comes before
`[ 'a', 1 ]` and `[ 'b', 1 ]` both, it's up next, but on the next
selection `[ 'a', 1 ]` comes before `[ 'b', 1 ]` so now we are out of
order.

The above description is from the commit log message before the release
of Skip 0.2.7. Link below.

The new `map` function of Homogenize will assume that all iterators have
the same key for each entry in the forward iteration order. It will
merge the items of each entry in the map and sort them.

https://github.com/bigeasy/skip/releases/tag/v0.2.7

Closes #108.