Remove outer key assertion from map.
It would require the user to provide an inner and outer comparator. The
inner key is going to be a compound versioned key and the outer key is
going to to be the first element. These are going to be different
comparators. We'll have to just trust that we're getting the same set
from each iterator we homogenize.

Closes #111.