Upgrade `advance` to `3.0.0-alpha.8`.
Instead of `.forward`, `.reverse` and `.map` there is now an iterator
type property that is used to determine the type of designator iterator
to create based on the type of the decorated iterator. We use constants
defined in the MVCC constants module `mvcc` to identify the type. The
Homogenize iterator will inherit the type of the decorated iterator.

https://github.com/bigeasy/mvcc/releases/tag/v2.0.0-alpha.2
https://github.com/bigeasy/advance/releases/tag/v3.0.0-alpha.8

Closes #113.