Upgrade `ascension` to `3.0.0-alpha.0`.
Ascension no longer has an extractor function. Instead we now use the
Whittle module to create a comparator decorator that whittles an
object down to a compound key for use with Ascension. We do not need to
use Whittle here, though. We're working directly with keys in our unit
tests and not with records that need a key extracted.

https://github.com/bigeasy/ascension/releases/tag/v3.0.0-alpha.0

Closes #115.