Sort iterators by length.
Need to sort the iterators by length in the map iterator for the generic
iterator advance function to work. The forward and reverse merge are
always sorting the iterator array to get the next version and add it to
the set of ordered, homogenized values. The furthest advanced is always
first and that allows it to detect when the index has reached the end.

Because all the mapped iterators are ultimately the same length we sort
by size to put the iterator that broke the set merge loop first so it
can be advanced.